+++ src/themes/_generator/tokens/zIndex/zIndex.transforms.ts
@@ -1,0 +1,13 @@
+import type * as T from "./zIndex.types";
+import type { Transformer } from "../types";
+
+const transformToken: Transformer<T.Token> = (name, token) => [
+	{
+		name,
+		tokenType: "zIndex",
+		type: "variable",
+		value: `${token.level}`,
+	},
+];
+
+export default transformToken;