+++ src/themes/_generator/tokens/zIndex/zIndex.types.ts
@@ -1,0 +1,2 @@
+export type Name = "relative" | "absolute" | "fixed";
+export type Token = { level: number };