+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added Attributes type to the package exports
 - Added `mono` font family to the themes
+- Added zIndex tokens to the themes, updated components to use new zIndex tokens
+- Added isolation: isolate to components for better zIndex handling
 
 - Image: added maxWidth property
 - Image: added aspectRatio property